Use CommonQuert for data and availability queries
